+++ app/stores/TransactionsStore.js
@@ -153,6 +153,10 @@ export default class TransactionsStore extends Store {
     return result ? result.transactions.length : 0;
   }
 
+  calculateTransactionFee = (from: string, to: string, amount: string) => (
+    this.api.calculateTransactionFee({ from, to, amount: new BigNumber(amount) })
+  );
+
   @action _refreshTransactionData = () => {
     if (this.stores.networkStatus.isConnected) {
       const allWallets = this.stores.wallets.all;